desktop_app/requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
face-recognition==1.3.0
psycopg2-binary==2.9.7
gfpgan==1.3.8
numpy
dlib
cmake
